Move template.py to templates/ dir
