+++ dp_creator_ii/app/dataset_panel.py
@@ -2,7 +2,7 @@ from shiny import ui, reactive, render
 
 from dp_creator_ii.utils.argparse_helpers import get_csv_contrib
 from dp_creator_ii.app.components.outputs import output_code_sample
-from dp_creator_ii.utils.template import make_privacy_unit_block
+from dp_creator_ii.utils.templates import make_privacy_unit_block
 
 
 def dataset_ui():
