+++ dp_creator_ii/app/results_panel.py
@@ -1,6 +1,6 @@
 from shiny import ui, render
 
-from dp_creator_ii.utils.template import make_notebook_py, make_script_py
+from dp_creator_ii.utils.templates import make_notebook_py, make_script_py
 from dp_creator_ii.utils.converters import convert_py_to_nb
